# Optimize Rlp.LengthOf(UInt256 item)