--- a/src/Paprika.Tests/Merkle/HashingTests.cs
+++ b/src/Paprika.Tests/Merkle/HashingTests.cs
@@ -52,7 +52,7 @@ public class HashingTests
         var expectedHash = new Keccak(Convert.FromHexString(hexString));
 
         Assert.That(computedHash.DataType, Is.EqualTo(KeccakOrRlp.Type.Keccak));
-        Assert.That(new Keccak(computedHash.Span), Is.EqualTo(expectedHash));
+        Assert.That(computedHash.Keccak, Is.EqualTo(expectedHash));
     }
 
     [Test]