### Benchmarks

Using `MerkleBenchmarks` for this

#### Before

| Method             | Mean      | Error    | StdDev   | Code Size | Allocated |
|------------------- |----------:|---------:|---------:|----------:|----------:|
| EOA_Leaf           | 280.77 ns | 1.803 ns | 1.598 ns |     890 B |         - |
| Contract_Leaf      | 281.04 ns | 4.823 ns | 4.511 ns |     890 B |         - |
| Storage_long_path  | 241.74 ns | 1.664 ns | 1.299 ns |     482 B |         - |
| Storage_short_path |  37.19 ns | 0.371 ns | 0.329 ns |     485 B |         - |

#### After

| Method             | Mean      | Error    | StdDev   | Code Size | Allocated |
|------------------- |----------:|---------:|---------:|----------:|----------:|
| EOA_Leaf           | 260.69 ns | 2.766 ns | 2.452 ns |     810 B |         - |
| Contract_Leaf      | 265.04 ns | 2.695 ns | 2.250 ns |     810 B |         - |
| Storage_long_path  | 230.28 ns | 1.270 ns | 1.188 ns |     469 B |         - |
| Storage_short_path |  27.24 ns | 0.197 ns | 0.184 ns |     472 B |         - |