@benaadams 2024-04-05
Should be better now; though bigger change sharing output buffer via `out` rather than return and not daisy chaining by `this` returns of large structs:

Before

| Method             | Mean      | Error    | StdDev   |
|------------------- |----------:|---------:|---------:|
| EOA_Leaf           | 276.24 ns | 2.072 ns | 1.939 ns |
| Contract_Leaf      | 275.52 ns | 2.487 ns | 2.205 ns |
| Storage_long_path  | 240.84 ns | 1.716 ns | 1.521 ns |
| Storage_short_path |  38.42 ns | 0.118 ns | 0.099 ns |

After

| Method             | Mean      | Error    | StdDev   |
|------------------- |----------:|---------:|---------:|
| EOA_Leaf           | 259.54 ns | 2.364 ns | 2.211 ns |
| Contract_Leaf      | 259.35 ns | 1.401 ns | 1.170 ns |
| Storage_long_path  | 222.39 ns | 1.068 ns | 0.892 ns |
| Storage_short_path |  23.95 ns | 0.070 ns | 0.059 ns |