@benaadams 2024-04-05
Though try locally to update summary since you have more columns 🤔