@benaadams 2024-04-05
| Method             | Mean      | Error    | StdDev   |
|------------------- |----------:|---------:|---------:|
| EOA_Leaf           | 257.25 ns | 0.907 ns | 0.708 ns |
| Contract_Leaf      | 255.70 ns | 0.836 ns | 0.698 ns |
| Storage_long_path  | 223.75 ns | 2.621 ns | 2.323 ns |
| Storage_short_path |  23.90 ns | 0.153 ns | 0.136 ns |